requirements.txt
pillow
imagehash==4.3.1
requests
